\name{wec-preprocess-package}
\alias{wec-preprocess-package}
\alias{wec-preprocess}
\docType{package}
\title{
\packageTitle{wec-preprocess}
}
\description{
\packageDescription{wec-preprocess}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wec-preprocess}
\packageIndices{wec-preprocess}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{wec-preprocess}

Maintainer: \packageMaintainer{wec-preprocess}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
